+++ apps/expo-go/modules/react-native-view-shot/src/RNViewShot.js
@@ -1,3 +1,0 @@
-//@flow
-import RNViewShot from './specs/NativeRNViewShot'
-export default RNViewShot